% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tlc}
\alias{tlc}
\title{Funcion TLC para la media}
\usage{
tlc(var1, n, k, t)
}
\arguments{
\item{var1}{Variable a utilizar}

\item{n}{Tama??o o tama??os de muestras}

\item{k}{n??mero de iteraciones}

\item{t}{tiempo de espera}
}
\description{
Funcion TLC para la media
}
\examples{
variable1=rnorm(1000,50,10);tlc(variable1,c(32,200),20,2)

}
