% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tlc_razon}
\alias{tlc_razon}
\title{Funcion TLC para la razon de dos variables}
\usage{
tlc_razon(x1, y1, n, k, t)
}
\arguments{
\item{x1}{Variable 1}

\item{y1}{Variable 2}

\item{n}{Tama??o o tama??os de muestras}

\item{k}{n??mero de iteraciones}

\item{t}{tiempo de espera}
}
\description{
Funcion TLC para la razon de dos variables
}
\examples{
variable1=rnorm(1000,50,10);Variable2=rpois(1000,60);n1=c(5,70,100) ;tlc_razon(variable1,Variable2,n1,10,2)

}
